ai/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
groq==0.4.2
